Create fields with QuerySets when forms are instantiated
Form fields which include a QuerySet as one of their arguments and are
defined on a class will have their QuerySet executed when the form is
imported.  These forms are imported by views which are in turn imported
by URLs meaning Django always tried to run the QuerySets so one is
unable to run even the migrations.  By defining the fields in __init__
methods the QuerySets aren't triggered until the form is instantiated.