Use Version to define published state for all related objects
Now that Databases have a version and all data is input via the
importers (instead of via User input) we can rely on a single versioned
import being published (or not) and derive all display logic from that.

This adds a middleware to set a version instance on all requests.  It
chooses the latest published Version unless the current User has a
current_version set.

All views now "hang" off request.version to define which items are
returned for the currently "active" version.