Convert publish views to Version and fix search
Now that Version handles published state the publish toggle and
unpublished list view need to handle Version.  As a knock on of this the
search view also needs to be updated to handle how we now look up
objects.  Since search and the dynamic model forms were so tied together
this also removes the generic form views & templates.