Centralise log output to respective stderr & stdout files
This removes the gunicorn config file since it was both a source of
configuration indirection and didn't allow the use of hyphen to redirect
to standard streams.  As part of that it also consolidates all output
from Gunicorn, Django, and the RQ worker into the same streams.

gunicorn.log is no longer created since logfile is synonymous with
error-logfile to gunicorn and we now redirect that to stderr.